pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "nodelab"
version = "0.1.0"
description = "TODO."
authors = [
    { name = "Sinan Kurtyigit", email = "sinan.kurtyigit@gmail.com" }
]
dependencies = [
    "pyaml"
]
